+++ tests/test_rules/express.ts
@@ -5,12 +5,16 @@ import passport from 'passport';
 function authenticated() {
     const router1 = Router()
     const router2 = express.Router()
+    const app = express()
 
     router1.use(passport.initialize());
     router1.use(passport.session());
 
     router2.use(passport.initialize());
     router2.use(passport.session());
+
+    app.use(passport.initialize());
+    app.use(passport.session());
 
     // ruleid: express-route-authenticated
     router1.get('/', (req, res) => {
@@ -32,11 +36,17 @@ function authenticated() {
     router2.delete('/', (req, res) => {
       res.send('DELETE request to the homepage')
     })
+
+    // ruleid: express-route-authenticated
+    app.get('/', (req, res) => {
+      res.send('hello world')
+    })
 }
 
 function unauthenticated() {
     const router1 = Router()
     const router2 = express.Router()
+    const app = express()
 
     // ruleid: express-route-unauthenticated
     router1.get('/', (req, res) => {
@@ -46,5 +56,10 @@ function unauthenticated() {
     // ruleid: express-route-unauthenticated
     router2.delete('/', (req, res) => {
       res.send('DELETE request to the homepage')
+    })
+
+    // ruleid: express-route-unauthenticated
+    app.get('/', (req, res) => {
+      res.send('hello world')
     })
 }