+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for Express routes defined on the app ([#16](https://github.com/mschwager/route-detect/issues/16))
+
 ## [0.7.0] - 2023-06-28
 
 ### Added